+++ src/library/adash-messages.adb
@@ -206,6 +206,8 @@ package body Adash.Messages is
             return "error.ambiguous_literal";
          when Msg_Raise_Outside_A_Handler =>
             return "error.raise_outside_a_handler";
+         when Msg_Aggregate_Others_Covers_Nothing =>
+            return "error.aggregate.others_covers_nothing";
          when Msg_Empty_Priority_Range =>
             return "error.empty_priority_range";
          when Msg_Restriction_Broken =>
@@ -993,7 +995,8 @@ package body Adash.Messages is
 
          when Msg_Select_Alternative | Msg_Select_Waits_Twice
             | Msg_Select_Trigger | Msg_Terminate_Outside_Select
-            | Msg_Raise_Outside_A_Handler =>
+            | Msg_Raise_Outside_A_Handler
+            | Msg_Aggregate_Others_Covers_Nothing =>
             return Placeholder_Names'(1 .. 0 => <>);
 
          when Msg_Discriminants_Need_A_Type | Msg_Nothing_To_Constrain